fix(export): double the reverse solidus in STEP string literals (write side of #2323)
ISO 10303-21 doubles two characters inside a STEP string literal: the
apostrophe and the reverse solidus. `escape()` in the STEP exporter doubled
the apostrophe but not the backslash, so a property/pset name containing a
literal `\` (a Windows path, a regex) was written as an under-escaped,
non-conformant literal — `C:\temp` stayed `C:\temp` instead of `C:\\temp`.
This is the write-side half of the doubling-escape gap; the read-side half
(decode_ifc_string collapsing `\\`) is tracked separately and is not yet on
main, so this fix is written to be correct both before and after that lands
(doubling on write is a spec requirement independent of what any reader does).

The fix alone pushes step.rs from 611 to 617 lines, past its module-size
ratchet budget. Rather than raise the budget, the text-level STEP primitives
that don't touch DATA-section emission (escape, detect_schema, refs_in_line,
split_top_level_args, apply_attr_mutations) are split out to step_text.rs —
a real seam, not a line-count split. step.rs settles back at exactly 471
lines post-split-and-fix, so the allowlist budget only moves down (611 -> 471),
never up. The move was verified behaviour-preserving by running the full
workspace suite before and after the split, as a step separate from the
escaping fix (1591 passed / 0 failed both times).

Caller audit: every escape() call site (header fields, property/pset names)
receives raw, never-escaped text; AttrMutation.value / PropMutation.value are
documented as already STEP-serialized and are spliced in raw, untouched by
escape() — so this fix cannot double-escape an already-escaped value.

Tests: unit-level RED/GREEN pin on escape() for the backslash arm (plus a
combined apostrophe+backslash case and a byte-identical bounding control for
plain strings), and an end-to-end test that synthesizes a property set with
both characters in its name and checks the raw written bytes against a
strict ISO 10303-21 un-escaper (a real reader is entitled to reject an
odd-length backslash run, which is exactly what the pre-fix output produced).
Mutation-tested: reverting the fix fails all three new tests.

Changeset added for @ifc-lite/geometry, which wraps the wasm `exportStep`
that this exporter backs.
